iconDefinitions home to icon defs, sorting, etc
